Add password hashing and validation functions, rename model file
