+++ backend/requirements.txt
@@ -6,3 +6,4 @@ requests==2.32.3
 pydantic[email]==2.10.6
 python-dotenv==1.0.1
 motor==3.7.0
+bcrypt==4.2.1
